Updated network scanning tasks to reflect scan config, new NetworkTask base task class.
